Refactor phase prompts and session management for improved workflow handling
This commit consolidates and refines the phase prompts and session management logic. It introduces new functions for enhanced session context resolution and dynamic input validation, ensuring a more robust workflow execution process. Additionally, it updates the YAML parsing and validation mechanisms, streamlining the handling of workflow definitions. The changes also include improved error handling and logging for better traceability during session operations.
